change Product quantity from fixedpoint to integer
- quantity column in Products sheet of test.xlsx now contains only integers
- quantity is stored as integer in order_lineitems
Jira: LIMS-1936

+++ bika/lims/browser/order.py
@@ -40,7 +40,7 @@ class View(BrowserView):
 		    product = [pro for pro in products if pro.getId() == prodid][0]
 		    price = float(item['Price'])
 		    vat = float(item['VAT'])
-		    qty = float(item['Quantity'])
+		    qty = item['Quantity']
 		    self.items.append({
 		        'title': product.Title(),
 		        'description': product.Description(),
@@ -56,6 +56,7 @@ class View(BrowserView):
 
 	def getPreferredCurrencyAbreviation(self):
 		return self.context.bika_setup.getCurrency()
+
 
 class EditView(BrowserView):
 
@@ -80,12 +81,12 @@ class EditView(BrowserView):
             context.order_lineitems = []
             # Process the order item data
             for prodid, qty in request.form.items():
-                if prodid.startswith('product_') and float(qty) > 0:
+                if prodid.startswith('product_') and int(qty) > 0:
                     prodid = prodid.replace('product_', '')
                     product = [pro for pro in products if pro.getId() == prodid][0]
                     context.order_lineitems.append(
                             {'Product': prodid,
-                             'Quantity': qty,
+                             'Quantity': int(qty),
                              'Price': product.getPrice(),
                              'VAT': product.getVAT()})
